fix decimal number

+++ js/bank.js
@@ -2,7 +2,7 @@ document.getElementById("deposit-btn").addEventListener("click", function () {
   // get the new deposit input value
   const depositInput = document.getElementById("deposit-input");
   console.log(typeof depositInput.value);
-  const newDeposit = parseInt(depositInput.value);
+  const newDeposit = parseFloat(depositInput.value);
   console.log(typeof newDeposit);
   if (depositInput.value == "") {
     return alert("please input some value");
@@ -15,13 +15,13 @@ document.getElementById("deposit-btn").addEventListener("click", function () {
 
   // get the old deposit value
   const oldDepositText = document.getElementById("deposit");
-  const oldDepositValue = parseInt(oldDepositText.innerText);
+  const oldDepositValue = parseFloat(oldDepositText.innerText);
 
   oldDepositText.innerText = oldDepositValue + newDeposit;
   depositInput.value = "";
   // add the deposit money in balance
   const oldBalanceText = document.getElementById("balance");
-  const oldBalanceValue = parseInt(oldBalanceText.innerText);
+  const oldBalanceValue = parseFloat(oldBalanceText.innerText);
   oldBalanceText.innerText = oldBalanceValue + newDeposit;
 });
 document.getElementById("withdraw-btn").addEventListener("click", function () {
@@ -35,17 +35,18 @@ document.getElementById("withdraw-btn").addEventListener("click", function () {
     withdrawInput.value = "";
     return;
   } */
-  const newWithdrawValue = parseInt(withdrawInput.value);
+
+  const newWithdrawValue = parseFloat(withdrawInput.value);
   // get the balance value
   const oldBalanceText = document.getElementById("balance");
-  const oldBalanceValue = parseInt(oldBalanceText.innerText);
+  const oldBalanceValue = parseFloat(oldBalanceText.innerText);
   // set the new value in balance
 
   oldBalanceText.innerText = oldBalanceValue - newWithdrawValue;
 
   // set the withdraw money in withdraw
   const oldWithdrawText = document.getElementById("withdraw");
-  const oldWithdrawValue = parseInt(oldWithdrawText.innerText);
+  const oldWithdrawValue = parseFloat(oldWithdrawText.innerText);
   oldWithdrawText.innerText = oldWithdrawValue + newWithdrawValue;
   withdrawInput.value = "";
 });
